backend/requirements.txt
fastapi
uvicorn
tensorflow
nltk
asyncpg
pandas
python-dotenv
python-multipart